Player Player Site v1
This is the first version of the Player Player landing site.
I developed this using terrible git practices. My first commit is
literally the first finished version of the project.

- The site is responsive.
- Illustrations used from https://illlustrations.co/

I need to go back and refactor a couple of components to make this
package as lightweight and DRY as possible.